[Refactor] Share the MoE grouped-GEMM nopad workload
The test and the benchmark each defined a MoeGroupedGemmNopadTest
subclassing WorkloadBase, with the same expert-size formula written twice.
Move input generation to workloads/moe.py as the test stage's MUST PROVIDE
artifact and have both stages import it; the benchmark's ref_program stub,
which returned None, goes with it.
