+++ workloads/moe.py
@@ -72,3 +72,72 @@ class MoeUnpermuteTest(WorkloadBase):
             self.total_tokens, self.top_k, dtype=torch.float32, device="cuda"
         )
         return mm2_pad, fwd_idx, topk_weights
+
+
+def make_expert_sizes_offsets(
+    numel: int,
+    num_experts: int,
+    distribution: str,
+    device: str,
+) -> tuple[torch.Tensor, torch.Tensor]:
+    """Build (true_sizes, true_offsets) for a fixed token-to-expert distribution.
+
+    Args:
+        numel: Total token-expert pairs (T * top_k).
+        num_experts: Number of experts E.
+        distribution: "uniform" — evenly split; "skewed" — most tokens on first
+            20% of experts (one-token floor for the rest).
+        device: CUDA device string.
+
+    Returns:
+        true_sizes [E] int32, true_offsets [E] int32.
+    """
+    if distribution == "uniform":
+        base = max(1, numel // num_experts)
+        sizes = torch.full((num_experts,), base, dtype=torch.int32, device=device)
+        sizes[-1] = numel - base * (num_experts - 1)
+    elif distribution == "skewed":
+        sizes = torch.ones(num_experts, dtype=torch.int32, device=device)
+        extra = numel - num_experts
+        top_experts = max(1, num_experts // 5)
+        per_top = extra // top_experts
+        sizes[:top_experts] += per_top
+        sizes[0] += extra - per_top * top_experts
+    else:
+        raise ValueError(f"unknown distribution: {distribution}")
+
+    offsets = torch.zeros(num_experts, dtype=torch.int32, device=device)
+    offsets[1:] = torch.cumsum(sizes[:-1], dim=0)
+    assert int(sizes.sum().item()) == numel
+    return sizes, offsets
+
+
+class MoeGroupedGemmNopadWorkload(WorkloadBase):
+    """Tight A, per-expert weights B, and the expert size/offset tables."""
+
+    def __init__(
+        self,
+        numel: int,
+        num_experts: int,
+        n: int,
+        k: int,
+        dtype: torch.dtype,
+        distribution: str = "uniform",
+    ):
+        self.numel = numel
+        self.num_experts = num_experts
+        self.n = n
+        self.k = k
+        self.dtype = dtype
+        self.distribution = distribution
+
+    def gen_inputs(self):
+        torch.manual_seed(42)
+        dev = "cuda"
+        true_sizes, true_offsets = make_expert_sizes_offsets(
+            self.numel, self.num_experts, self.distribution, dev
+        )
+        # Small scale keeps fp16 accumulation well within the parity tolerance.
+        a = torch.randn(self.numel, self.k, dtype=self.dtype, device=dev) * 0.02
+        b = torch.randn(self.num_experts, self.n, self.k, dtype=self.dtype, device=dev) * 0.02
+        return a, b, true_sizes, true_offsets
